pyproject.toml
[project]
name = "academical"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord-py>=2.5.2",
    "fastapi>=0.116.0",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "uvicorn>=0.35.0",
]
